backend/requirements.txt
flask
flask-cors
numpy
pillow
tensorflow
werkzeug
requests
